requirements.txt
Django==1.7.1
Pillow==2.5.1
PyYAML==3.11
argh==0.25.0
django-admin-sortable2==0.3.1
django-appconf==0.6
django-bower==5.0.1
django-braces==1.4.0
django-ckeditor==4.4.6
django-classy-tags==0.5.1
django-compressor==1.4
django-endless-pagination==2.0
django-extensions==1.3.8
django-lineage==0.2.0
django-sekizai==0.7
django-shell-plus==1.1.5
django-supervisor==0.3.2
django-taggit==0.12.2
django-taggit-templatetags==0.2.5
django-tastypie==0.12.1
django-templatetag-sugar==1.0
gunicorn==19.0.0
meld3==1.0.0
pathtools==0.1.2
psycopg2==2.5.3
python-dateutil==2.3
python-mimeparse==0.1.4
pytz==2014.4
six==1.7.3
supervisor==3.0
watchdog==0.8.0
wsgiref==0.1.2
